# Make internal dict iterator structs not allocate.